lifeCounter veranderd in stepCounter
zodat er op degene die het verst is wordt gemikt

+++ Tower Defense/Machine.java
@@ -89,7 +89,7 @@ public class Machine extends Actor
             if(type != 4) {
                 ArrayList counterArrayInRange = new ArrayList(robotsInRange.size());
                 for(Robot a : (List<Robot>) robotsInRange) {
-                    int counterCounter = a.getCounter();
+                    double counterCounter = a.getCounter();
                     counterArrayInRange.add(counterCounter);
                 }
